+++ glyph/src/ttfp/outliner.rs
@@ -63,3 +63,33 @@ impl owned_ttf_parser::OutlineBuilder for OutlineCurveBuilder {
         }
     }
 }
+
+/// Compatibility struct for ttf-parser using a OutlineBuilder.
+pub(crate) struct CompatOutlineBuilder<'a>(pub &'a mut dyn crate::OutlineBuilder);
+
+impl owned_ttf_parser::OutlineBuilder for CompatOutlineBuilder<'_> {
+    #[inline]
+    fn move_to(&mut self, x: f32, y: f32) {
+        self.0.move_to(point(x, y));
+    }
+
+    #[inline]
+    fn line_to(&mut self, x: f32, y: f32) {
+        self.0.line_to(point(x, y));
+    }
+
+    #[inline]
+    fn quad_to(&mut self, x1: f32, y1: f32, x2: f32, y2: f32) {
+        self.0.quad_to(point(x1, y1), point(x2, y2));
+    }
+
+    #[inline]
+    fn curve_to(&mut self, x1: f32, y1: f32, x2: f32, y2: f32, x3: f32, y3: f32) {
+        self.0.curve_to(point(x1, y1), point(x2, y2), point(x3, y3))
+    }
+
+    #[inline]
+    fn close(&mut self) {
+        self.0.close();
+    }
+}